Decompose CLI
