+++ plusdeck/cli/logger.py
@@ -1,0 +1,9 @@
+from typing import Literal
+
+LogLevel = (
+    Literal["DEBUG"]
+    | Literal["INFO"]
+    | Literal["WARNING"]
+    | Literal["ERROR"]
+    | Literal["CRITICAL"]
+)
